+++ src/pygaps/parsing/json.py
@@ -4,15 +4,15 @@ import json
 
 import pandas
 
-from ..core.isotherm import Isotherm
-from ..core.modelisotherm import ModelIsotherm
-from ..core.pointisotherm import PointIsotherm
-from ..modelling import get_isotherm_model
-from ..utilities.exceptions import ParsingError
-from ..utilities.unit_converter import _MASS_UNITS
-from ..utilities.unit_converter import _MOLAR_UNITS
-from ..utilities.unit_converter import _PRESSURE_UNITS
-from ..utilities.unit_converter import _VOLUME_UNITS
+from pygaps.core.isotherm import Isotherm
+from pygaps.core.modelisotherm import ModelIsotherm
+from pygaps.core.pointisotherm import PointIsotherm
+from pygaps.modelling import model_from_dict
+from pygaps.utilities.exceptions import ParsingError
+from pygaps.utilities.unit_converter import _MASS_UNITS
+from pygaps.utilities.unit_converter import _MOLAR_UNITS
+from pygaps.utilities.unit_converter import _PRESSURE_UNITS
+from pygaps.utilities.unit_converter import _VOLUME_UNITS
 
 
 def isotherm_to_json(isotherm, path=None, **args_to_json):
@@ -164,31 +164,11 @@ def isotherm_from_json(
             **raw_dict
         )
     elif model:
-
-        new_mod = get_isotherm_model(model['model'])
-
-        rmse = model.pop('rmse', None)
-        if rmse:
-            new_mod.rmse = rmse
-
-        prange = model.pop('pressure_range', None)
-        if prange:
-            new_mod.pressure_range = prange
-
-        lrange = model.pop('loading_range', None)
-        if lrange:
-            new_mod.loading_range = lrange
-
-        for param in new_mod.params:
-            try:
-                new_mod.params[param] = model['parameters'][param]
-            except KeyError as err:
-                raise KeyError(
-                    f"The JSON is missing parameter '{param}'"
-                ) from err
-
         # generate the isotherm
-        isotherm = ModelIsotherm(model=new_mod, **raw_dict)
+        isotherm = ModelIsotherm(
+            model=model_from_dict(model),
+            **raw_dict,
+        )
 
     else:
         # generate the isotherm